requirements.txt
# Core parsing libraries
PyPDF2==3.0.1              # PDF text extraction
python-docx==1.1.0         # Word document parsing

# NLP for name extraction
spacy==3.7.2               # NER for name extraction
# Run after install: python -m spacy download en_core_web_sm

# LLM integration
openai==1.6.1              # GPT-4 API
tiktoken==0.5.2            # Token counting for cost estimation

# Configuration & environment
python-dotenv==1.0.0       # Environment variable management
pydantic==2.5.3            # Data validation
pydantic-settings==2.1.0   # Settings management

# Utilities
tenacity==8.2.3            # Retry logic for API calls

